requirements.txt
flask>=2.0
requests>=2.25
python-dotenv>=0.20
markdown>=3.4
